Rename package `query.internal` to `query.core` in soil-query-core
