+++ soil-query-core/src/androidMain/kotlin/soil/query/core/Platform.android.kt
@@ -1,7 +1,7 @@
 // Copyright 2024 Soil Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package soil.query.internal
+package soil.query.core
 
 import java.util.UUID
