+++ soil-query-core/src/commonMain/kotlin/soil/query/QueryCommand.kt
@@ -3,13 +3,13 @@
 
 package soil.query
 
-import soil.query.internal.RetryCallback
-import soil.query.internal.RetryFn
-import soil.query.internal.UniqueId
-import soil.query.internal.epoch
-import soil.query.internal.exponentialBackOff
-import soil.query.internal.toEpoch
-import soil.query.internal.vvv
+import soil.query.core.RetryCallback
+import soil.query.core.RetryFn
+import soil.query.core.UniqueId
+import soil.query.core.epoch
+import soil.query.core.exponentialBackOff
+import soil.query.core.toEpoch
+import soil.query.core.vvv
 import kotlin.coroutines.cancellation.CancellationException
 
 /**
