+++ soil-query-core/src/commonMain/kotlin/soil/query/QueryOptions.kt
@@ -3,12 +3,12 @@
 
 package soil.query
 
-import soil.query.internal.ActorOptions
-import soil.query.internal.LoggerFn
-import soil.query.internal.LoggingOptions
-import soil.query.internal.RetryOptions
-import soil.query.internal.Retryable
-import soil.query.internal.UniqueId
+import soil.query.core.ActorOptions
+import soil.query.core.LoggerFn
+import soil.query.core.LoggingOptions
+import soil.query.core.RetryOptions
+import soil.query.core.Retryable
+import soil.query.core.UniqueId
 import kotlin.random.Random
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
@@ -49,7 +49,7 @@ interface QueryOptions : ActorOptions, LoggingOptions, RetryOptions {
      * Automatically revalidate active [Query] when the network reconnects.
      *
      * **Note:**
-     * This setting is only effective when [soil.query.internal.NetworkConnectivity] is available.
+     * This setting is only effective when [soil.query.core.NetworkConnectivity] is available.
      */
     val revalidateOnReconnect: Boolean
 
@@ -57,7 +57,7 @@ interface QueryOptions : ActorOptions, LoggingOptions, RetryOptions {
      * Automatically revalidate active [Query] when the window is refocused.
      *
      * **Note:**
-     * This setting is only effective when [soil.query.internal.WindowVisibility] is available.
+     * This setting is only effective when [soil.query.core.WindowVisibility] is available.
      */
     val revalidateOnFocus: Boolean
