+++ soil-query-core/src/commonMain/kotlin/soil/query/SwrCache.kt
@@ -28,20 +28,20 @@ import kotlinx.coroutines.withContext
 import kotlinx.coroutines.withTimeoutOrNull
 import soil.query.SwrCachePolicy.Companion.DEFAULT_GC_CHUNK_SIZE
 import soil.query.SwrCachePolicy.Companion.DEFAULT_GC_INTERVAL
-import soil.query.internal.ActorBlockRunner
-import soil.query.internal.ActorSequenceNumber
-import soil.query.internal.MemoryPressure
-import soil.query.internal.MemoryPressureLevel
-import soil.query.internal.NetworkConnectivity
-import soil.query.internal.NetworkConnectivityEvent
-import soil.query.internal.SurrogateKey
-import soil.query.internal.TimeBasedCache
-import soil.query.internal.UniqueId
-import soil.query.internal.WindowVisibility
-import soil.query.internal.WindowVisibilityEvent
-import soil.query.internal.chunkedWithTimeout
-import soil.query.internal.epoch
-import soil.query.internal.vvv
+import soil.query.core.ActorBlockRunner
+import soil.query.core.ActorSequenceNumber
+import soil.query.core.MemoryPressure
+import soil.query.core.MemoryPressureLevel
+import soil.query.core.NetworkConnectivity
+import soil.query.core.NetworkConnectivityEvent
+import soil.query.core.SurrogateKey
+import soil.query.core.TimeBasedCache
+import soil.query.core.UniqueId
+import soil.query.core.WindowVisibility
+import soil.query.core.WindowVisibilityEvent
+import soil.query.core.chunkedWithTimeout
+import soil.query.core.epoch
+import soil.query.core.vvv
 import kotlin.coroutines.CoroutineContext
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
