+++ soil-query-core/src/commonMain/kotlin/soil/query/core/NetworkOptions.kt
@@ -1,7 +1,7 @@
 // Copyright 2024 Soil Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package soil.query.internal
+package soil.query.core
 
 interface NetworkOptions {
     val isNetworkError: ((Throwable) -> Boolean)?
