+++ soil-query-core/src/wasmJsMain/kotlin/soil/query/core/Platform.wasmJs.kt
@@ -1,7 +1,7 @@
 // Copyright 2024 Soil Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package soil.query.internal
+package soil.query.core
 
 actual fun epoch(): Long {
     return Date.now().toString().toLong() / 1000
